LayoutTests/fast/dom/document-all-tags.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>
description('Test that HTMLAllCollection.tags() matches IE behavior with too few arguments.');
shouldThrow("document.all.tags()");
successfullyParsed = true;
</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
